Order va Order_itemsdagi buglarni tuzatdim

+++ models/order_items.py
@@ -1,5 +1,5 @@
 # models/order_item.py
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, Float, ForeignKey
 from sqlalchemy.orm import relationship
 from db import Base
 
@@ -10,7 +10,7 @@ class OrderItem(Base):
     id = Column(Integer, autoincrement=True, primary_key=True)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
     amount = Column(Integer, nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(Float, nullable=False)
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
 
     product = relationship("Products", back_populates="order_items")
